python/requirements.txt
python-uinput
pyserial
